Make CI green, and fix a SyntaxError on the default install path
CI was red on every push and would have stayed red through the public flip.
Three separate causes, none of them visible from a passing local test run.

1. hub/inspect.py was a SyntaxError on Python 3.11.

   It used `\x00` escapes inside f-string EXPRESSIONS, which is 3.12 syntax.
   requires-python is >=3.11, there is no .python-version, and `uv sync` picks
   3.11 -- so a fresh clone got 3.11 and the module failed to import. That
   module is what the data-read guard tells every agent to use instead of
   `cat`, so on a clean install the recommended escape hatch was broken.

   The separator is now a named constant. Verified: all 95 files parse under
   3.11, and the full suite passes on 3.11.12.

2. The `golden` job had never run, once, ever.

   It gates on `github.event_name == 'schedule'` in a workflow triggered
   `on: [push, pull_request]`. Its own comment called it nightly and the README
   said the same. The trigger is now present and commented out until the flip,
   matching watchdog.yml's metering rationale, and the flip checklist names it.

3. `ruff check src tests` reported 186 errors.

   [tool.ruff] set only line-length, so the lint contract was whatever the
   installed ruff version defaulted to -- 186 errors under 0.16 against 31
   under the documented E4/E7/E9/F. A gate whose strictness changes when a
   dependency updates is not a gate. Now an explicit select, and clean.

   The valuable part was B905: 24 `zip()` calls without `strict=`. Adding
   strict=True surfaced three real length mismatches -- all the
   `zip(xs, xs[1:])` pairwise idiom, now `itertools.pairwise`. The rest are
   genuinely equal-length and now say so.

   DTZ deliberately not selected: 26 naive datetime calls are real, but fixing
   them changes stored timestamps and that is not freeze-week work.

Coverage was 78.18% against CI's 80% floor, because the harnesses I added this
session had untested orchestration -- which is the exact failure hub.draft.
backtest exists to prevent. compare, play, diagnose and both strategies are now
exercised offline. 80.24%.

Also rewrote README.md, which described a system that does not exist:
hierarchical Bayesian NumPyro ratings (ratings.py is an explicit passthrough
whose own docstring says it has no edge), MAPIE Mondrian conformal (mapie is a
declared dependency imported nowhere), CLV (not implemented), and nightly
golden tests (see 2). For a repo whose stated purpose is method under
adversarial self-audit, an overclaiming front door is the worst possible one --
and the true story is stronger, so it now leads with the five things that were
built, measured and removed.

849 passing. ruff, pyrefly and the coverage floor all green.

+++ src/hub/models/components.py
@@ -27,7 +27,7 @@ fitted coefficient of variation.
 """
 from __future__ import annotations
 
-from typing import Iterable, Mapping
+from collections.abc import Iterable, Mapping
 
 import numpy as np
 import polars as pl
@@ -112,7 +112,7 @@ def _counts(rng: np.random.Generator, mean: float, phi: float, n: int) -> np.nda
         # var/mean = 1 - p for a binomial, so p is fixed by the dispersion and the number
         # of trials by the mean.
         p = 1.0 - phi
-        trials = int(round(mean / p))
+        trials = round(mean / p)
         return rng.binomial(max(trials, 1), min(mean / max(trials, 1), 1.0), n)
     return rng.poisson(mean, n)
